Add Web Server
- fix warnings and update action versions
- Web Server added in example
- prepare release 0.7.5
