+++ library.properties
@@ -1,9 +1,12 @@
 name=DShotRMT
-version=0.7.2
-author=derdoktor667
-maintainer=derdoktor667
+version=0.7.5
+author=Wastl Kraus <wir-sind-die-matrix.de>
+maintainer=Wastl Kraus <wir-sind-die-matrix.de>
+license=MIT
 sentence=DShotRMT Library supporting all DShot Types and speeds. Tested with BlHeli_S.
-paragraph=This library can control a BlHeli_S by using encoded DShot commands.
+paragraph=This library can control a BlHeli_S by using encoded DShot commands. Features bidirectional DShot support for RPM telemetry.
 category=Signal Input/Output
 url=https://github.com/derdoktor667/DShotRMT
 architectures=esp32
+provides_includes=DShotRMT.h, DShotCommandManager.h, dshot_commands.h, web_content.h
+depends=ArduinoJson
