+++ src/web_content.h
@@ -6,6 +6,8 @@
  * @license MIT
  */
 
+#pragma once
+
 // Web Site Content
 const char index_html[] PROGMEM = R"rawliteral(
 <!DOCTYPE html>
